infra/ocr/requirements.txt
# --- core pins (NumPy 1.x for OpenCV<=4.6) ---
numpy==1.26.4

fastapi==0.118.0
uvicorn[standard]==0.37.0
pillow==11.3.0
paddlepaddle==2.6.1

# headless OpenCV to avoid libGL
opencv-python-headless==4.6.0.66
opencv-contrib-python-headless==4.6.0.66

# paddleocr runtime deps
shapely==2.1.2
scikit-image==0.25.2
imgaug==0.4.0
pyclipper==1.3.0.post6
lmdb==1.7.3
tqdm==4.67.1
visualdl==2.5.3
rapidfuzz==3.14.1
PyMuPDF<1.21.0
pyyaml==6.0.3
python-docx==1.2.0
beautifulsoup4==4.14.2
lxml==6.0.2
premailer==3.10.0
openpyxl==3.1.5
attrdict==2.0.1
pdf2docx==0.5.8

# required by FastAPI for file uploads
python-multipart==0.0.20
